Replace old CI script
(cherry picked from commit 390c1e7def0eef6d39ca64a0d11165be45684257)
